+++ SocketExe/Server/Server.cpp
@@ -37,7 +37,7 @@ int main(int argc, char *argv[]){
     std::string from_ip;
     from_ip.resize(255);
     std::array<char, 65535> buffer;
-    ssize_t read_bytes;
+    long long read_bytes;
     read_bytes = Socket::recvFrom(sockfd, from_ip, buffer);
     if (read_bytes < 0) {
         Socket::close(sockfd);
